requirements.txt
notion-client
openai
python-dotenv